+++ lib/live_activities_method_channel.dart
@@ -1,13 +1,14 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+import 'package:live_activities/live_activities_platform_interface.dart';
 import 'package:live_activities/models/activity_update.dart';
+import 'package:live_activities/models/alert_config.dart';
 import 'package:live_activities/models/live_activity_state.dart';
 import 'package:live_activities/models/url_scheme_data.dart';
-
-import 'live_activities_platform_interface.dart';
 
 /// An implementation of [LiveActivitiesPlatform] that uses method channels.
 class MethodChannelLiveActivities extends LiveActivitiesPlatform {
@@ -16,17 +17,19 @@ class MethodChannelLiveActivities extends LiveActivitiesPlatform {
   final methodChannel = const MethodChannel('live_activities');
 
   @visibleForTesting
-  final activityStatusChannel =
-      const EventChannel('live_activities/activity_status');
+  final activityStatusChannel = const EventChannel('live_activities/activity_status');
 
   @visibleForTesting
-  final EventChannel urlSchemeChannel =
-      const EventChannel('live_activities/url_scheme');
+  final EventChannel urlSchemeChannel = const EventChannel('live_activities/url_scheme');
+
+  @visibleForTesting
+  final EventChannel pushToStartTokenUpdatesChannel = const EventChannel('live_activities/push_to_start_token_updates');
 
   @override
-  Future init(String appGroupId) async {
+  Future init(String appGroupId, {String? urlScheme}) async {
     await methodChannel.invokeMethod('init', {
       'appGroupId': appGroupId,
+      'urlScheme': urlScheme,
     });
   }
 
@@ -37,8 +40,7 @@ class MethodChannelLiveActivities extends LiveActivitiesPlatform {
     Duration? staleIn,
   }) async {
     // If the duration is less than 1 minute then pass a null value instead of using 0 minutes
-    final staleInMinutes =
-        (staleIn?.inMinutes ?? 0) >= 1 ? staleIn?.inMinutes : null;
+    final staleInMinutes = (staleIn?.inMinutes ?? 0) >= 1 ? staleIn?.inMinutes : null;
     return methodChannel.invokeMethod<String>('createActivity', {
       'data': data,
       'removeWhenAppIsKilled': removeWhenAppIsKilled,
@@ -47,10 +49,24 @@ class MethodChannelLiveActivities extends LiveActivitiesPlatform {
   }
 
   @override
-  Future updateActivity(String activityId, Map<String, dynamic> data) async {
-    return methodChannel.invokeMethod('updateActivity', {
-      'activityId': activityId,
+  Future updateActivity(String activityId, Map<String, dynamic> data, [AlertConfig? alertConfig]) async {
+    return methodChannel
+        .invokeMethod('updateActivity', {'activityId': activityId, 'data': data, 'alertConfig': alertConfig?.toMap()});
+  }
+
+  @override
+  Future createOrUpdateActivity(
+    String customId,
+    Map<String, dynamic> data, {
+    bool removeWhenAppIsKilled = false,
+    Duration? staleIn,
+  }) async {
+    final staleInMinutes = (staleIn?.inMinutes ?? 0) >= 1 ? staleIn?.inMinutes : null;
+    return methodChannel.invokeMethod('createOrUpdateActivity', {
+      'customId': customId,
       'data': data,
+      'removeWhenAppIsKilled': removeWhenAppIsKilled,
+      'staleIn': staleInMinutes,
     });
   }
 
@@ -68,35 +84,54 @@ class MethodChannelLiveActivities extends LiveActivitiesPlatform {
 
   @override
   Future<List<String>> getAllActivitiesIds() async {
-    final result =
-        await methodChannel.invokeListMethod<String>('getAllActivitiesIds');
+    final result = await methodChannel.invokeListMethod<String>('getAllActivitiesIds');
     return result ?? [];
   }
 
   @override
+  Future<Map<String, LiveActivityState>> getAllActivities() async {
+    final result = await methodChannel.invokeMapMethod<String, String>('getAllActivities');
+
+    return result?.map((key, value) => MapEntry(key, LiveActivityState.values.byName(value))) ??
+        <String, LiveActivityState>{};
+  }
+
+  @override
   Future<bool> areActivitiesEnabled() async {
-    final result =
-        await methodChannel.invokeMethod<bool>('areActivitiesEnabled');
+    if (!Platform.isIOS) {
+      return false;
+    }
+
+    final result = await methodChannel.invokeMethod<bool>('areActivitiesEnabled');
+    return result ?? false;
+  }
+
+  @override
+  Future<bool> allowsPushStart() async {
+    if (!Platform.isIOS) {
+      return false;
+    }
+
+    final result = await methodChannel.invokeMethod<bool>('allowsPushStart');
     return result ?? false;
   }
 
   @override
   Stream<UrlSchemeData> urlSchemeStream() {
     return urlSchemeChannel.receiveBroadcastStream('urlSchemeStream').map(
-          (dynamic event) =>
-              UrlSchemeData.fromMap(Map<String, dynamic>.from(event)),
+          (dynamic event) => UrlSchemeData.fromMap(Map<String, dynamic>.from(event)),
         );
   }
 
   @override
-  Future<LiveActivityState> getActivityState(String activityId) async {
-    final result = await methodChannel.invokeMethod<String>(
+  Future<LiveActivityState?> getActivityState(String activityId) async {
+    final result = await methodChannel.invokeMethod<String?>(
       'getActivityState',
       {
         'activityId': activityId,
       },
     );
-    return LiveActivityState.values.byName(result ?? 'unknown');
+    return result != null ? LiveActivityState.values.byName(result) : null;
   }
 
   @override
@@ -115,4 +150,8 @@ class MethodChannelLiveActivities extends LiveActivitiesPlatform {
       .receiveBroadcastStream('activityUpdateStream')
       .distinct()
       .map((event) => ActivityUpdate.fromMap(Map<String, dynamic>.from(event)));
+
+  @override
+  Stream<String> get pushToStartTokenUpdateStream =>
+      pushToStartTokenUpdatesChannel.receiveBroadcastStream('pushToStartTokenUpdateStream').distinct().cast<String>();
 }