+++ test/live_activities_test.dart
@@ -3,6 +3,7 @@ import 'package:live_activities/live_activities.dart';
 import 'package:live_activities/live_activities_platform_interface.dart';
 import 'package:live_activities/live_activities_method_channel.dart';
 import 'package:live_activities/models/activity_update.dart';
+import 'package:live_activities/models/alert_config.dart';
 import 'package:live_activities/models/live_activity_state.dart';
 import 'package:live_activities/models/url_scheme_data.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -11,7 +12,7 @@ class MockLiveActivitiesPlatform
     with MockPlatformInterfaceMixin
     implements LiveActivitiesPlatform {
   @override
-  Future init(String appGroupId) {
+  Future init(String appGroupId, {String? urlScheme}) {
     return Future.value();
   }
 
@@ -26,11 +27,6 @@ class MockLiveActivitiesPlatform
 
   @override
   Future endActivity(String activityId) {
-    return Future.value();
-  }
-
-  @override
-  Future updateActivity(String activityId, Map<String, dynamic> data) {
     return Future.value();
   }
 
@@ -81,6 +77,37 @@ class MockLiveActivitiesPlatform
     };
     return Stream.value(ActivityUpdate.fromMap(map));
   }
+
+  @override
+  Future updateActivity(String activityId, Map<String, dynamic> data,
+      [AlertConfig? alertConfig]) {
+    return Future.value();
+  }
+
+  @override
+  Future<Map<String, LiveActivityState>> getAllActivities() {
+    return Future.value({'ACTIVITY_ID': LiveActivityState.active});
+  }
+
+  @override
+  Future createOrUpdateActivity(
+    String customId,
+    Map<String, dynamic> data, {
+    bool removeWhenAppIsKilled = false,
+    Duration? staleIn,
+  }) {
+    return Future.value();
+  }
+
+  @override
+  Future<bool> allowsPushStart() {
+    return Future.value(true);
+  }
+
+  @override
+  Stream<String> get pushToStartTokenUpdateStream {
+    return Stream.value('PUSH_TO_START_TOKEN');
+  }
 }
 
 void main() {
@@ -110,8 +137,14 @@ void main() {
     expect(await liveActivitiesPlugin.endAllActivities(), null);
   });
 
-  test('getAllActivities', () async {
+  test('getAllActivitiesIds', () async {
     expect(await liveActivitiesPlugin.getAllActivitiesIds(), ['ACTIVITY_ID']);
+  });
+
+  test('getAllActivities', () async {
+    expect(await liveActivitiesPlugin.getAllActivities(), {
+      'ACTIVITY_ID': LiveActivityState.active,
+    });
   });
 
   test('areActivitiesEnabled', () async {
@@ -166,5 +199,16 @@ void main() {
         result.mapOrNull(active: (state) => state.activityToken);
 
     expect(correctMappingNotNull, 'ACTIVITY_TOKEN');
+  });
+
+  test('allowsPushStart', () async {
+    expect(await liveActivitiesPlugin.allowsPushStart(), true);
+  });
+
+  test('pushToStartTokenUpdateStream', () async {
+    expect(
+      await liveActivitiesPlugin.pushToStartTokenUpdateStream.first,
+      'PUSH_TO_START_TOKEN',
+    );
   });
 }