+++ CHANGELOG.md
@@ -1,3 +1,124 @@
+## 2.3.1
+* 🐛 `LiveActivityFileFromMemory` can't share image with AppGroup (thanks to @EArminjon 👍).
+* 📝 Added minor version check when not being able to see the Live Activity (thanks to @dasanten 👍).
+* ⬆️ Upgrade dependencies.
+
+## 2.3.0
+* 🏗️ Move to Swift Package Manager.
+* 🏗️ Regenerate example app.
+* ⬆️ Upgrade dependencies.
+
+## 2.2.0
+* ✨ Added a new method `createOrUpdateActivity()`, you can use it to create or update by passing an activity ID (thanks to @Clon1998 👍).
+* ⬆️ Upgrade dependencies.
+
+## 2.1.0
+* ✨ You can now send generic files instead of just pictures.
+  
+To send a file you can do the following on the Dart code:
+
+```dart
+LiveActivityFileFromAsset('assets/files/rules.txt')
+```
+
+And in your Swift code:
+
+```swift
+let ruleFile = sharedDefault.string(forKey: context.attributes.prefixedKey("yourFileKey"))!
+let rule = (try? String(contentsOfFile: ruleFile, encoding: .utf8)) ?? ""
+```
+
+> Check the example for a full demo.
+
+**BREAKING CHANGES**
+
+You must replace your images to:
+- `LiveActivityFileFromAsset.image()`
+- `LiveActivityFileFromMemory.image()`
+- `LiveActivityFileFromUrl.image()`
+
+In order to use `resizeFactor` you must do the following now:
+
+```dart
+LiveActivityFileFromAsset.image(
+  'assets/images/chelsea.png',
+  imageOptions: LiveActivityImageFileOptions(
+    resizeFactor: 0.2
+  )
+),
+```
+
+* 🐛 Example works again with scheme.
+
+## 2.0.1
+* 🐛 Fix channel message sent from native to Flutter on a non-platform thread (thanks to @aswanath 👍)
+* 🗑️ Clean some code.
+
+## 2.0.0
+* ✨ Use new custom iOS [App Group Directory dependency](https://pub.dev/packages/flutter_app_group_directory) (this will fix namespace for Android gradle builds).
+* ✨ Removed deprecated flutter_native_image dependency and replaced by image (thanks to @SnapDrive 👍)
+* ⬆️ Upgrade dependencies.
+
+## 1.9.5
+* 🐛 Fix `areActivitiesEnabled()` on unsupported devices.
+
+## 1.9.4
+* 🍱 Convert images to webp.
+* ⬆️ Upgrade dependencies.
+
+## 1.9.3
+* 🐛 Force returning false for `areActivitiesEnabled()` when no iOS devices.
+
+## 1.9.2
+* ✨ Simplified fetching of ActivityState of all created live activities (thanks to @Clon1998 👍).
+* 🐛 Fixes background thread invocation of event streams (thanks to @ggirotto 👍).
+* 🐛 Replaced getImageProperties with dart buffer and descriptor (thanks to @anumb 👍).
+* 🐛 Fix tests.
+* ⬆️ Upgrade dependencies.
+
+## 1.9.1
+
+* ✨ Add update with alert config (thanks @charlesRmajor 👍).
+* ✨ Add an option to use preloaded images (thanks @Niklas-Sommer 👍).
+* ✨ Add Android support - currently only used to check if live activities is supported (thanks @ggirotto 👍).
+* ✨ Example app support Material 3.
+* 🐛 Fix tests.
+* 📝 Update README.md.
+* ⬆️ Upgrade dependencies.
+
+## 1.9.0
+
+- ✨ **BREAKING CHANGE**: Add the ability to handle multiple live notification (thanks @Clon1998 👍).
+
+Please follow this tutorial to add implement it:
+
+- Add the following Swift extension at the end of your extension code:
+
+```swift
+extension LiveActivitiesAppAttributes {
+  func prefixedKey(_ key: String) -> String {
+    return "\(id)_\(key)"
+  }
+}
+```
+
+- For each keys on your native Swift code, please changes the following lines:
+
+```swift
+let myVariableFromFlutter = sharedDefault.string(forKey: "myVariableFromFlutter") // repleace this by ...
+let myVariableFromFlutter = sharedDefault.string(forKey: context.attributes.prefixedKey("myVariableFromFlutter")) // <-- this
+```
+ 
+- 🐛 Fix stall state for unknown activityId (thanks @Clon1998 👍).
+- 🐛 Now return `null` value when activity is not found in `getActivityState()`.
+
+## 1.8.0
+
+* ✨ Add url scheme optional argument.
+* ✨ Add sinks unregister on engine end (thanks @ggirotto 👍).
+* 🐛 Fix example images size.
+* ⬆️ Upgrade dependencies.
+
 ## 1.7.5
 
 * 🚨 Lint some code.